Put every test in one file because of multiple browser creation. Modified README.md. Introduced Background to .feature file.
