+++ pom.xml
@@ -20,6 +20,7 @@
         <maven.compiler.version>3.3</maven.compiler.version>
         <selenium.version>2.53.0</selenium.version>
         <surefire.version>2.16</surefire.version>
+        <firefox.folder>"Default\path\to\folder"</firefox.folder>
     </properties>
 
     <dependencies>
@@ -65,6 +66,11 @@
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-surefire-plugin</artifactId>
                 <version>${surefire.version}</version>
+                <configuration>
+                    <systemPropertyVariables>
+                        <firefox.folder>//Default//folder//</firefox.folder>
+                    </systemPropertyVariables>
+                </configuration>
             </plugin>
         </plugins>
     </build>
