+++ src/test/java/com/balazspap/docler/stepdef/TestRunner.java
@@ -1,4 +1,4 @@
-package com.balazspap.docler;
+package com.balazspap.docler.stepdef;
 
 import cucumber.api.CucumberOptions;
 import cucumber.api.junit.Cucumber;
@@ -9,6 +9,7 @@ import org.junit.runner.RunWith;
         plugin = {"pretty", "html:target/cucumber-html-report"},
         format = {"json:target/integration_cucumber.json"},
         features="src/test/resources"
+//        glue = {"src/test/java/com/balazspap/docler/stepdef"}
 )
 public class TestRunner {
 }
