--- src/test/resources/logoButtonFunctionality.feature
@@ -1,9 +1,0 @@
-Feature: Logo button functionality test
-  Background:
-  Requirements:
-  REQ-UI-08: When I click on the Logo, I should get navigated to the Home page (High Priority)
-
-  Scenario: When I click on the Logo, I should get navigated to the Home page
-    Given I get to UI Testing home page
-    When I click the Logo button
-    Then I get directed to the Home page
